+++ src/models/PanelSettings.ts
@@ -48,7 +48,7 @@ export interface ContentType {
   pageBundle?: boolean;
 }
 
-export type FieldType = "string" | "number" | "datetime" | "boolean" | "image" | "choice" | "tags" | "categories" | "draft" | "taxonomy" | "fields" | "json" | "block" | "file" | "dataFile" | "list";
+export type FieldType = "string" | "number" | "datetime" | "boolean" | "image" | "choice" | "tags" | "categories" | "draft" | "taxonomy" | "fields" | "json" | "block" | "file" | "dataFile" | "list" | "slug";
 
 export interface Field {
   title?: string;
@@ -67,6 +67,7 @@ export interface Field {
   dataType?: string | string[];
   taxonomyLimit?: number;
   fileExtensions?: string[];
+  editable?: boolean;
 
   // Date fields
   isPublishDate?: boolean;